requirements.txt
grpcio==1.76.0
grpcio-tools==1.76.0
python-dotenv==1.2.1
requests==2.32.5
httpx==0.28.1